% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{add_file}
\alias{add_file}
\title{Add files to metagenomicsClientR's cache}
\usage{
add_file(rname, url)
}
\arguments{
\item{rname}{A character vector indicating the resource name (file).}

\item{url}{A character vector indicating the url of the resource}
}
\value{
Character vector of legnth 1, indicating resource name and path
}
\description{
\code{add_file} adds (or downloads) a file to metagenomicsClientR's cache.
}
\details{
This function implements BiocFileCache::bfcadd to easily add files to the metagenomicsClientR's cache. If the file is already present in the cache or needs to be updated, \code{add_file} prints a message and returns the path to the file.
}
\examples{
 
}
