% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nf_events.R
\name{check_null}
\alias{check_null}
\title{Check if value is NULL}
\usage{
check_null(x)
}
\arguments{
\item{x}{an object to be tested}
}
\value{
logical
}
\description{
Helper function for get_nf_events()
}
