% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_api.R
\docType{class}
\name{DefaultApi}
\alias{DefaultApi}
\title{Default operations}
\format{
An \code{R6Class} generator object
}
\description{
metagenomicsClientR.Default
}
\section{Methods}{

\describe{
\strong{ AddNextflowEventNextflowEventsPost } \emph{ Add Nextflow Event }
Events from running Nextflow pipelines when using -with-weblog.  See [the Nextflow documentation](https://www.nextflow.io/docs/latest/tracing.html#weblog-via-http).

\itemize{
\item \emph{ @param } nf.model \link{NFModel}
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ CreateSampleMetadataSamplesPost } \emph{ Create Sample }


\itemize{
\item \emph{ @param } sample \link{Sample}
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ CreateStudyMetadataStudiesPost } \emph{ Create Study }


\itemize{
\item \emph{ @param } study \link{Study}
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ DeleteStudyMetadataStudiesUuidDelete } \emph{ Delete Study }


\itemize{
\item \emph{ @param } uuid \link{character}
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ FilesChangeFilesChangesPost } \emph{ Files Change }
Track all changes in cloud storage

\itemize{
\item \emph{ @param } body object
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ GenerateDownloadSignedUrlV4FilesSignedGet } \emph{ Generate Download Signed Url V4 }
Generates a v4 signed URL for downloading a blob.  Note that this method requires a service account key file. You can not use this if you are using Application Default Credentials from Google Compute Engine or from the Google Cloud SDK.

\itemize{
\item \emph{ @param } name character
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetNextflowEventNextflowEventsIdGet } \emph{ Get Nextflow Event }


\itemize{
\item \emph{ @param } id \link{character}
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetSamplesMetadataSamplesGet } \emph{ Get Samples }


\itemize{
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetStudiesMetadataStudiesGet } \emph{ Get Studies }


\itemize{
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetUuidFromStringMetadataUuidGet } \emph{ Get Uuid From String }


\itemize{
\item \emph{ @param } string character
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ ListFileEventsFilesChangesGet } \emph{ List File Events }


\itemize{
\item \emph{ @param } limit integer
\item \emph{ @param } offset integer
\item \emph{ @param } start.date character
\item \emph{ @param } end.date character
\item \emph{ @param } filename.regex character
\item \emph{ @returnType } \link{StorageEventCollection} \cr


\item status code : 200 | Successful Response

\item return type : StorageEventCollection 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ ListNextflowEventsNextflowEventsGet } \emph{ List Nextflow Events }


\itemize{
\item \emph{ @param } limit integer
\item \emph{ @param } offset integer
\item \emph{ @returnType } \link{NFCollection} \cr


\item status code : 200 | Successful Response

\item return type : NFCollection 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ SamplesForStudyMetadataStudiesUuidSamplesGet } \emph{ Samples For Study }


\itemize{
\item \emph{ @param } uuid \link{character}
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ SamplesJsonMetadataSamplesjsonGet } \emph{ Samples Json }


\itemize{
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  AddNextflowEventNextflowEventsPost  ####################

library(metagenomicsClientR)
var.nf.model <- NFModel$new() # NFModel | 

#Add Nextflow Event
api.instance <- DefaultApi$new()

result <- api.instance$AddNextflowEventNextflowEventsPost(var.nf.model)


####################  CreateSampleMetadataSamplesPost  ####################

library(metagenomicsClientR)
var.sample <- Sample$new() # Sample | 

#Create Sample
api.instance <- DefaultApi$new()

result <- api.instance$CreateSampleMetadataSamplesPost(var.sample)


####################  CreateStudyMetadataStudiesPost  ####################

library(metagenomicsClientR)
var.study <- Study$new() # Study | 

#Create Study
api.instance <- DefaultApi$new()

result <- api.instance$CreateStudyMetadataStudiesPost(var.study)


####################  DeleteStudyMetadataStudiesUuidDelete  ####################

library(metagenomicsClientR)
var.uuid <- 'uuid_example' # character | 

#Delete Study
api.instance <- DefaultApi$new()

result <- api.instance$DeleteStudyMetadataStudiesUuidDelete(var.uuid)


####################  FilesChangeFilesChangesPost  ####################

library(metagenomicsClientR)
var.body <- NULL # object | 

#Files Change
api.instance <- DefaultApi$new()

result <- api.instance$FilesChangeFilesChangesPost(var.body)


####################  GenerateDownloadSignedUrlV4FilesSignedGet  ####################

library(metagenomicsClientR)
var.name <- 'name_example' # character | 

#Generate Download Signed Url V4
api.instance <- DefaultApi$new()

result <- api.instance$GenerateDownloadSignedUrlV4FilesSignedGet(var.name)


####################  GetNextflowEventNextflowEventsIdGet  ####################

library(metagenomicsClientR)
var.id <- 'id_example' # character | 

#Get Nextflow Event
api.instance <- DefaultApi$new()

result <- api.instance$GetNextflowEventNextflowEventsIdGet(var.id)


####################  GetSamplesMetadataSamplesGet  ####################

library(metagenomicsClientR)

#Get Samples
api.instance <- DefaultApi$new()

result <- api.instance$GetSamplesMetadataSamplesGet()


####################  GetStudiesMetadataStudiesGet  ####################

library(metagenomicsClientR)

#Get Studies
api.instance <- DefaultApi$new()

result <- api.instance$GetStudiesMetadataStudiesGet()


####################  GetUuidFromStringMetadataUuidGet  ####################

library(metagenomicsClientR)
var.string <- 'string_example' # character | 

#Get Uuid From String
api.instance <- DefaultApi$new()

result <- api.instance$GetUuidFromStringMetadataUuidGet(var.string)


####################  ListFileEventsFilesChangesGet  ####################

library(metagenomicsClientR)
var.limit <- 100 # integer | 
var.offset <- 0 # integer | 
var.start.date <- 'start.date_example' # character | 
var.end.date <- 'end.date_example' # character | 
var.filename.regex <- 'filename.regex_example' # character | 

#List File Events
api.instance <- DefaultApi$new()

result <- api.instance$ListFileEventsFilesChangesGet(limit=var.limit, offset=var.offset, start.date=var.start.date, end.date=var.end.date, filename.regex=var.filename.regex)


####################  ListNextflowEventsNextflowEventsGet  ####################

library(metagenomicsClientR)
var.limit <- 100 # integer | 
var.offset <- 0 # integer | 

#List Nextflow Events
api.instance <- DefaultApi$new()

result <- api.instance$ListNextflowEventsNextflowEventsGet(limit=var.limit, offset=var.offset)


####################  SamplesForStudyMetadataStudiesUuidSamplesGet  ####################

library(metagenomicsClientR)
var.uuid <- 'uuid_example' # character | 

#Samples For Study
api.instance <- DefaultApi$new()

result <- api.instance$SamplesForStudyMetadataStudiesUuidSamplesGet(var.uuid)


####################  SamplesJsonMetadataSamplesjsonGet  ####################

library(metagenomicsClientR)

#Samples Json
api.instance <- DefaultApi$new()

result <- api.instance$SamplesJsonMetadataSamplesjsonGet()


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DefaultApi$new()}}
\item \href{#method-AddNextflowEventNextflowEventsPost}{\code{DefaultApi$AddNextflowEventNextflowEventsPost()}}
\item \href{#method-AddNextflowEventNextflowEventsPostWithHttpInfo}{\code{DefaultApi$AddNextflowEventNextflowEventsPostWithHttpInfo()}}
\item \href{#method-CreateSampleMetadataSamplesPost}{\code{DefaultApi$CreateSampleMetadataSamplesPost()}}
\item \href{#method-CreateSampleMetadataSamplesPostWithHttpInfo}{\code{DefaultApi$CreateSampleMetadataSamplesPostWithHttpInfo()}}
\item \href{#method-CreateStudyMetadataStudiesPost}{\code{DefaultApi$CreateStudyMetadataStudiesPost()}}
\item \href{#method-CreateStudyMetadataStudiesPostWithHttpInfo}{\code{DefaultApi$CreateStudyMetadataStudiesPostWithHttpInfo()}}
\item \href{#method-DeleteStudyMetadataStudiesUuidDelete}{\code{DefaultApi$DeleteStudyMetadataStudiesUuidDelete()}}
\item \href{#method-DeleteStudyMetadataStudiesUuidDeleteWithHttpInfo}{\code{DefaultApi$DeleteStudyMetadataStudiesUuidDeleteWithHttpInfo()}}
\item \href{#method-FilesChangeFilesChangesPost}{\code{DefaultApi$FilesChangeFilesChangesPost()}}
\item \href{#method-FilesChangeFilesChangesPostWithHttpInfo}{\code{DefaultApi$FilesChangeFilesChangesPostWithHttpInfo()}}
\item \href{#method-GenerateDownloadSignedUrlV4FilesSignedGet}{\code{DefaultApi$GenerateDownloadSignedUrlV4FilesSignedGet()}}
\item \href{#method-GenerateDownloadSignedUrlV4FilesSignedGetWithHttpInfo}{\code{DefaultApi$GenerateDownloadSignedUrlV4FilesSignedGetWithHttpInfo()}}
\item \href{#method-GetNextflowEventNextflowEventsIdGet}{\code{DefaultApi$GetNextflowEventNextflowEventsIdGet()}}
\item \href{#method-GetNextflowEventNextflowEventsIdGetWithHttpInfo}{\code{DefaultApi$GetNextflowEventNextflowEventsIdGetWithHttpInfo()}}
\item \href{#method-GetSamplesMetadataSamplesGet}{\code{DefaultApi$GetSamplesMetadataSamplesGet()}}
\item \href{#method-GetSamplesMetadataSamplesGetWithHttpInfo}{\code{DefaultApi$GetSamplesMetadataSamplesGetWithHttpInfo()}}
\item \href{#method-GetStudiesMetadataStudiesGet}{\code{DefaultApi$GetStudiesMetadataStudiesGet()}}
\item \href{#method-GetStudiesMetadataStudiesGetWithHttpInfo}{\code{DefaultApi$GetStudiesMetadataStudiesGetWithHttpInfo()}}
\item \href{#method-GetUuidFromStringMetadataUuidGet}{\code{DefaultApi$GetUuidFromStringMetadataUuidGet()}}
\item \href{#method-GetUuidFromStringMetadataUuidGetWithHttpInfo}{\code{DefaultApi$GetUuidFromStringMetadataUuidGetWithHttpInfo()}}
\item \href{#method-ListFileEventsFilesChangesGet}{\code{DefaultApi$ListFileEventsFilesChangesGet()}}
\item \href{#method-ListFileEventsFilesChangesGetWithHttpInfo}{\code{DefaultApi$ListFileEventsFilesChangesGetWithHttpInfo()}}
\item \href{#method-ListNextflowEventsNextflowEventsGet}{\code{DefaultApi$ListNextflowEventsNextflowEventsGet()}}
\item \href{#method-ListNextflowEventsNextflowEventsGetWithHttpInfo}{\code{DefaultApi$ListNextflowEventsNextflowEventsGetWithHttpInfo()}}
\item \href{#method-SamplesForStudyMetadataStudiesUuidSamplesGet}{\code{DefaultApi$SamplesForStudyMetadataStudiesUuidSamplesGet()}}
\item \href{#method-SamplesForStudyMetadataStudiesUuidSamplesGetWithHttpInfo}{\code{DefaultApi$SamplesForStudyMetadataStudiesUuidSamplesGetWithHttpInfo()}}
\item \href{#method-SamplesJsonMetadataSamplesjsonGet}{\code{DefaultApi$SamplesJsonMetadataSamplesjsonGet()}}
\item \href{#method-SamplesJsonMetadataSamplesjsonGetWithHttpInfo}{\code{DefaultApi$SamplesJsonMetadataSamplesjsonGetWithHttpInfo()}}
\item \href{#method-clone}{\code{DefaultApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddNextflowEventNextflowEventsPost"></a>}}
\if{latex}{\out{\hypertarget{method-AddNextflowEventNextflowEventsPost}{}}}
\subsection{Method \code{AddNextflowEventNextflowEventsPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$AddNextflowEventNextflowEventsPost(nf.model, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddNextflowEventNextflowEventsPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AddNextflowEventNextflowEventsPostWithHttpInfo}{}}}
\subsection{Method \code{AddNextflowEventNextflowEventsPostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$AddNextflowEventNextflowEventsPostWithHttpInfo(nf.model, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateSampleMetadataSamplesPost"></a>}}
\if{latex}{\out{\hypertarget{method-CreateSampleMetadataSamplesPost}{}}}
\subsection{Method \code{CreateSampleMetadataSamplesPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$CreateSampleMetadataSamplesPost(sample, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateSampleMetadataSamplesPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-CreateSampleMetadataSamplesPostWithHttpInfo}{}}}
\subsection{Method \code{CreateSampleMetadataSamplesPostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$CreateSampleMetadataSamplesPostWithHttpInfo(sample, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateStudyMetadataStudiesPost"></a>}}
\if{latex}{\out{\hypertarget{method-CreateStudyMetadataStudiesPost}{}}}
\subsection{Method \code{CreateStudyMetadataStudiesPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$CreateStudyMetadataStudiesPost(study, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CreateStudyMetadataStudiesPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-CreateStudyMetadataStudiesPostWithHttpInfo}{}}}
\subsection{Method \code{CreateStudyMetadataStudiesPostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$CreateStudyMetadataStudiesPostWithHttpInfo(study, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteStudyMetadataStudiesUuidDelete"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteStudyMetadataStudiesUuidDelete}{}}}
\subsection{Method \code{DeleteStudyMetadataStudiesUuidDelete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$DeleteStudyMetadataStudiesUuidDelete(uuid, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteStudyMetadataStudiesUuidDeleteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteStudyMetadataStudiesUuidDeleteWithHttpInfo}{}}}
\subsection{Method \code{DeleteStudyMetadataStudiesUuidDeleteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$DeleteStudyMetadataStudiesUuidDeleteWithHttpInfo(uuid, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilesChangeFilesChangesPost"></a>}}
\if{latex}{\out{\hypertarget{method-FilesChangeFilesChangesPost}{}}}
\subsection{Method \code{FilesChangeFilesChangesPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$FilesChangeFilesChangesPost(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilesChangeFilesChangesPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-FilesChangeFilesChangesPostWithHttpInfo}{}}}
\subsection{Method \code{FilesChangeFilesChangesPostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$FilesChangeFilesChangesPostWithHttpInfo(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenerateDownloadSignedUrlV4FilesSignedGet"></a>}}
\if{latex}{\out{\hypertarget{method-GenerateDownloadSignedUrlV4FilesSignedGet}{}}}
\subsection{Method \code{GenerateDownloadSignedUrlV4FilesSignedGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GenerateDownloadSignedUrlV4FilesSignedGet(name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenerateDownloadSignedUrlV4FilesSignedGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GenerateDownloadSignedUrlV4FilesSignedGetWithHttpInfo}{}}}
\subsection{Method \code{GenerateDownloadSignedUrlV4FilesSignedGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GenerateDownloadSignedUrlV4FilesSignedGetWithHttpInfo(name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetNextflowEventNextflowEventsIdGet"></a>}}
\if{latex}{\out{\hypertarget{method-GetNextflowEventNextflowEventsIdGet}{}}}
\subsection{Method \code{GetNextflowEventNextflowEventsIdGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetNextflowEventNextflowEventsIdGet(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetNextflowEventNextflowEventsIdGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetNextflowEventNextflowEventsIdGetWithHttpInfo}{}}}
\subsection{Method \code{GetNextflowEventNextflowEventsIdGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetNextflowEventNextflowEventsIdGetWithHttpInfo(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSamplesMetadataSamplesGet"></a>}}
\if{latex}{\out{\hypertarget{method-GetSamplesMetadataSamplesGet}{}}}
\subsection{Method \code{GetSamplesMetadataSamplesGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetSamplesMetadataSamplesGet(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetSamplesMetadataSamplesGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetSamplesMetadataSamplesGetWithHttpInfo}{}}}
\subsection{Method \code{GetSamplesMetadataSamplesGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetSamplesMetadataSamplesGetWithHttpInfo(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStudiesMetadataStudiesGet"></a>}}
\if{latex}{\out{\hypertarget{method-GetStudiesMetadataStudiesGet}{}}}
\subsection{Method \code{GetStudiesMetadataStudiesGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetStudiesMetadataStudiesGet(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetStudiesMetadataStudiesGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetStudiesMetadataStudiesGetWithHttpInfo}{}}}
\subsection{Method \code{GetStudiesMetadataStudiesGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetStudiesMetadataStudiesGetWithHttpInfo(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUuidFromStringMetadataUuidGet"></a>}}
\if{latex}{\out{\hypertarget{method-GetUuidFromStringMetadataUuidGet}{}}}
\subsection{Method \code{GetUuidFromStringMetadataUuidGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetUuidFromStringMetadataUuidGet(string, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUuidFromStringMetadataUuidGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetUuidFromStringMetadataUuidGetWithHttpInfo}{}}}
\subsection{Method \code{GetUuidFromStringMetadataUuidGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetUuidFromStringMetadataUuidGetWithHttpInfo(string, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFileEventsFilesChangesGet"></a>}}
\if{latex}{\out{\hypertarget{method-ListFileEventsFilesChangesGet}{}}}
\subsection{Method \code{ListFileEventsFilesChangesGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$ListFileEventsFilesChangesGet(
  limit = 100,
  offset = 0,
  start.date = NULL,
  end.date = NULL,
  filename.regex = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFileEventsFilesChangesGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ListFileEventsFilesChangesGetWithHttpInfo}{}}}
\subsection{Method \code{ListFileEventsFilesChangesGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$ListFileEventsFilesChangesGetWithHttpInfo(
  limit = 100,
  offset = 0,
  start.date = NULL,
  end.date = NULL,
  filename.regex = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListNextflowEventsNextflowEventsGet"></a>}}
\if{latex}{\out{\hypertarget{method-ListNextflowEventsNextflowEventsGet}{}}}
\subsection{Method \code{ListNextflowEventsNextflowEventsGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$ListNextflowEventsNextflowEventsGet(limit = 100, offset = 0, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListNextflowEventsNextflowEventsGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ListNextflowEventsNextflowEventsGetWithHttpInfo}{}}}
\subsection{Method \code{ListNextflowEventsNextflowEventsGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$ListNextflowEventsNextflowEventsGetWithHttpInfo(
  limit = 100,
  offset = 0,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SamplesForStudyMetadataStudiesUuidSamplesGet"></a>}}
\if{latex}{\out{\hypertarget{method-SamplesForStudyMetadataStudiesUuidSamplesGet}{}}}
\subsection{Method \code{SamplesForStudyMetadataStudiesUuidSamplesGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$SamplesForStudyMetadataStudiesUuidSamplesGet(uuid, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SamplesForStudyMetadataStudiesUuidSamplesGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SamplesForStudyMetadataStudiesUuidSamplesGetWithHttpInfo}{}}}
\subsection{Method \code{SamplesForStudyMetadataStudiesUuidSamplesGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$SamplesForStudyMetadataStudiesUuidSamplesGetWithHttpInfo(uuid, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SamplesJsonMetadataSamplesjsonGet"></a>}}
\if{latex}{\out{\hypertarget{method-SamplesJsonMetadataSamplesjsonGet}{}}}
\subsection{Method \code{SamplesJsonMetadataSamplesjsonGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$SamplesJsonMetadataSamplesjsonGet(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SamplesJsonMetadataSamplesjsonGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-SamplesJsonMetadataSamplesjsonGetWithHttpInfo}{}}}
\subsection{Method \code{SamplesJsonMetadataSamplesjsonGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$SamplesJsonMetadataSamplesjsonGetWithHttpInfo(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
