% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{download_file}
\alias{download_file}
\title{Add files to metagenomicsClientR}
\usage{
download_file(rname, url, verbose = FALSE)
}
\arguments{
\item{rname}{A character vector indicating the resource name (file).}

\item{url}{A character vector indicating the url of the resource}

\item{verbose}{logical}
}
\value{
Character vector of legnth 1, indicating resource name and path
}
\description{
This function wraps bcadd from BiocFileCache
}
\examples{



}
