% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_changes.R
\name{get_file_changes}
\alias{get_file_changes}
\title{Get list of file changes as tibble}
\usage{
get_file_changes(
  limit = 100,
  offset = 0,
  start_date = NULL,
  end_date = NULL,
  filename_regex = NULL
)
}
\arguments{
\item{limit}{A numeric vector of length 1 indicating the number of rows that must be returned.}

\item{offset}{TODO}

\item{start_date}{TODO}

\item{end_date}{TODO}

\item{filename_regex}{TODO}
}
\value{
tibble
}
\description{
TODO
}
\examples{
# Get the first 10 file changes
x <- get_file_changes(limit = 10)
x

}
