% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_client.R
\docType{data}
\name{metagenomics_client}
\alias{metagenomics_client}
\title{DefaultApi for metagenomicsClientR.}
\format{
An object of class \code{DefaultApi} (inherits from \code{R6}) of length 32.
}
\usage{
metagenomics_client
}
\description{
DefaultApi for metagenomicsClientR.
}
\keyword{datasets}
