% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomicsClientR.R
\docType{package}
\name{metagenomicsClientR}
\alias{metagenomicsClientR}
\title{metagenomicsCientR: R Package Client for Curated Metagenomics Web Portal and API}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{metagenomicsClientR functions}{

The foo functions ...
}

