% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextflow_event_model.R
\docType{class}
\name{NextflowEventModel}
\alias{NextflowEventModel}
\title{NextflowEventModel}
\format{
An \code{R6Class} generator object
}
\description{
NextflowEventModel Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{runName}}{Runname character}

\item{\code{runId}}{Runid character}

\item{\code{event}}{Event character}

\item{\code{utcTime}}{Utctime character}

\item{\code{trace}}{Trace object [optional]}

\item{\code{metadata}}{Metadata object [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NextflowEventModel$new()}}
\item \href{#method-toJSON}{\code{NextflowEventModel$toJSON()}}
\item \href{#method-fromJSON}{\code{NextflowEventModel$fromJSON()}}
\item \href{#method-toJSONString}{\code{NextflowEventModel$toJSONString()}}
\item \href{#method-fromJSONString}{\code{NextflowEventModel$fromJSONString()}}
\item \href{#method-clone}{\code{NextflowEventModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NextflowEventModel$new(
  runName,
  runId,
  event,
  utcTime,
  trace = NULL,
  metadata = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NextflowEventModel$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NextflowEventModel$fromJSON(NextflowEventModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NextflowEventModel$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NextflowEventModel$fromJSONString(NextflowEventModelJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NextflowEventModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
