% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nf_return.R
\docType{class}
\name{NFReturn}
\alias{NFReturn}
\title{NFReturn}
\format{
An \code{R6Class} generator object
}
\description{
NFReturn Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{run_name}}{Run Name character}

\item{\code{run_id}}{Run Id character}

\item{\code{event}}{Event character}

\item{\code{utc_time}}{Utc Time character}

\item{\code{trace}}{Trace object [optional]}

\item{\code{metadata}}{Metadata object [optional]}

\item{\code{event_id}}{Event Id integer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NFReturn$new()}}
\item \href{#method-toJSON}{\code{NFReturn$toJSON()}}
\item \href{#method-fromJSON}{\code{NFReturn$fromJSON()}}
\item \href{#method-toJSONString}{\code{NFReturn$toJSONString()}}
\item \href{#method-fromJSONString}{\code{NFReturn$fromJSONString()}}
\item \href{#method-clone}{\code{NFReturn$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NFReturn$new(
  run_name,
  run_id,
  event,
  utc_time,
  event_id,
  trace = NULL,
  metadata = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NFReturn$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NFReturn$fromJSON(NFReturnJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NFReturn$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NFReturn$fromJSONString(NFReturnJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NFReturn$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
