% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_from_dir.R
\name{read_file}
\alias{read_file}
\title{Read a single tsv file}
\usage{
read_file(x)
}
\arguments{
\item{x}{String indicating full path to a tsv file.}
}
\value{
tibble
}
\description{
Helper function for the \code{matrix_from_dir} function.
}
\examples{
 

}
