% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_event_return.R
\docType{class}
\name{StorageEventReturn}
\alias{StorageEventReturn}
\title{StorageEventReturn}
\format{
An \code{R6Class} generator object
}
\description{
StorageEventReturn Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{kind}}{Kind character}

\item{\code{id}}{Id character}

\item{\code{selfLink}}{Selflink character}

\item{\code{name}}{Name character}

\item{\code{bucket}}{Bucket character}

\item{\code{generation}}{Generation integer}

\item{\code{metageneration}}{Metageneration integer}

\item{\code{contentType}}{Contenttype character}

\item{\code{timeCreated}}{Timecreated character}

\item{\code{updated}}{Updated character}

\item{\code{size}}{Size integer}

\item{\code{md5Hash}}{Md5Hash character}

\item{\code{mediaLink}}{Medialink character}

\item{\code{contentLanguage}}{Contentlanguage character [optional]}

\item{\code{crc32c}}{Crc32C character}

\item{\code{etag}}{Etag character}

\item{\code{eventType}}{Eventtype character}

\item{\code{eventTime}}{Eventtime character}

\item{\code{event_id}}{Event Id character}

\item{\code{download_url}}{Download Url character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StorageEventReturn$new()}}
\item \href{#method-toJSON}{\code{StorageEventReturn$toJSON()}}
\item \href{#method-fromJSON}{\code{StorageEventReturn$fromJSON()}}
\item \href{#method-toJSONString}{\code{StorageEventReturn$toJSONString()}}
\item \href{#method-fromJSONString}{\code{StorageEventReturn$fromJSONString()}}
\item \href{#method-clone}{\code{StorageEventReturn$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageEventReturn$new(
  kind,
  id,
  selfLink,
  name,
  bucket,
  generation,
  metageneration,
  contentType,
  timeCreated,
  updated,
  size,
  md5Hash,
  mediaLink,
  crc32c,
  etag,
  eventType,
  eventTime,
  event_id,
  download_url,
  contentLanguage = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageEventReturn$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageEventReturn$fromJSON(StorageEventReturnJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageEventReturn$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageEventReturn$fromJSONString(StorageEventReturnJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageEventReturn$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
