% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_api.R
\docType{class}
\name{DefaultApi}
\alias{DefaultApi}
\title{Default operations}
\format{
An \code{R6Class} generator object
}
\description{
metagenomicsClientR.Default
}
\section{Methods}{

\describe{
\strong{ AddNextflowEventNextflowEventsPost } \emph{ Add Nextflow Event }
Events from running Nextflow pipelines when using -with-weblog.  See [the Nextflow documentation](https://www.nextflow.io/docs/latest/tracing.html#weblog-via-http).

\itemize{
\item \emph{ @param } nextflow.event.model \link{NextflowEventModel}
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ AddUserUsersPost } \emph{ Add User }


\itemize{
\item \emph{ @param } user.model \link{UserModel}
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ DeleteUserUsersUidDelete } \emph{ Delete User }


\itemize{
\item \emph{ @param } uid integer
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ FilesChangeFilesChangesPost } \emph{ Files Change }
Track all changes in cloud storage

\itemize{
\item \emph{ @param } body object
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ GenerateDownloadSignedUrlV4FilesSignedGet } \emph{ Generate Download Signed Url V4 }
Generates a v4 signed URL for downloading a blob.  Note that this method requires a service account key file. You can not use this if you are using Application Default Credentials from Google Compute Engine or from the Google Cloud SDK.

\itemize{
\item \emph{ @param } name character
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetNextflowEventNextflowEventsIdGet } \emph{ Get Nextflow Event }


\itemize{
\item \emph{ @param } id \link{character}
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ GetUserUsersUidGet } \emph{ Get User }


\itemize{
\item \emph{ @param } uid integer
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ ListFileEventsFilesChangesGet } \emph{ List File Events }


\itemize{
\item \emph{ @param } limit integer
\item \emph{ @param } offset integer
\item \emph{ @param } start.date character
\item \emph{ @param } end.date character
\item \emph{ @param } filename.regex character
\item \emph{ @returnType } \link{StorageEventCollection} \cr


\item status code : 200 | Successful Response

\item return type : StorageEventCollection 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ ListNextflowEventsNextflowEventsGet } \emph{ List Nextflow Events }


\itemize{
\item \emph{ @param } limit integer
\item \emph{ @param } offset integer
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
\item status code : 422 | Validation Error

\item return type : HTTPValidationError 
\item response headers :

\tabular{ll}{
}
}

\strong{ ListUsersUsersGet } \emph{ List Users }


\itemize{
\item \emph{ @returnType } \link{AnyType} \cr


\item status code : 200 | Successful Response

\item return type : AnyType 
\item response headers :

\tabular{ll}{
}
}

}
}

\examples{
\dontrun{
####################  AddNextflowEventNextflowEventsPost  ####################

library(metagenomicsClientR)
var.nextflow.event.model <- NextflowEventModel$new() # NextflowEventModel | 

#Add Nextflow Event
api.instance <- DefaultApi$new()

result <- api.instance$AddNextflowEventNextflowEventsPost(var.nextflow.event.model)


####################  AddUserUsersPost  ####################

library(metagenomicsClientR)
var.user.model <- UserModel$new() # UserModel | 

#Add User
api.instance <- DefaultApi$new()

result <- api.instance$AddUserUsersPost(var.user.model)


####################  DeleteUserUsersUidDelete  ####################

library(metagenomicsClientR)
var.uid <- 56 # integer | 

#Delete User
api.instance <- DefaultApi$new()

result <- api.instance$DeleteUserUsersUidDelete(var.uid)


####################  FilesChangeFilesChangesPost  ####################

library(metagenomicsClientR)
var.body <- NULL # object | 

#Files Change
api.instance <- DefaultApi$new()

result <- api.instance$FilesChangeFilesChangesPost(var.body)


####################  GenerateDownloadSignedUrlV4FilesSignedGet  ####################

library(metagenomicsClientR)
var.name <- 'name_example' # character | 

#Generate Download Signed Url V4
api.instance <- DefaultApi$new()

result <- api.instance$GenerateDownloadSignedUrlV4FilesSignedGet(var.name)


####################  GetNextflowEventNextflowEventsIdGet  ####################

library(metagenomicsClientR)
var.id <- 'id_example' # character | 

#Get Nextflow Event
api.instance <- DefaultApi$new()

result <- api.instance$GetNextflowEventNextflowEventsIdGet(var.id)


####################  GetUserUsersUidGet  ####################

library(metagenomicsClientR)
var.uid <- 56 # integer | 

#Get User
api.instance <- DefaultApi$new()

result <- api.instance$GetUserUsersUidGet(var.uid)


####################  ListFileEventsFilesChangesGet  ####################

library(metagenomicsClientR)
var.limit <- 100 # integer | 
var.offset <- 0 # integer | 
var.start.date <- 'start.date_example' # character | 
var.end.date <- 'end.date_example' # character | 
var.filename.regex <- 'filename.regex_example' # character | 

#List File Events
api.instance <- DefaultApi$new()

result <- api.instance$ListFileEventsFilesChangesGet(limit=var.limit, offset=var.offset, start.date=var.start.date, end.date=var.end.date, filename.regex=var.filename.regex)


####################  ListNextflowEventsNextflowEventsGet  ####################

library(metagenomicsClientR)
var.limit <- 100 # integer | 
var.offset <- 0 # integer | 

#List Nextflow Events
api.instance <- DefaultApi$new()

result <- api.instance$ListNextflowEventsNextflowEventsGet(limit=var.limit, offset=var.offset)


####################  ListUsersUsersGet  ####################

library(metagenomicsClientR)

#List Users
api.instance <- DefaultApi$new()

result <- api.instance$ListUsersUsersGet()


}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{apiClient}}{Handles the client-server communication.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DefaultApi$new()}}
\item \href{#method-AddNextflowEventNextflowEventsPost}{\code{DefaultApi$AddNextflowEventNextflowEventsPost()}}
\item \href{#method-AddNextflowEventNextflowEventsPostWithHttpInfo}{\code{DefaultApi$AddNextflowEventNextflowEventsPostWithHttpInfo()}}
\item \href{#method-AddUserUsersPost}{\code{DefaultApi$AddUserUsersPost()}}
\item \href{#method-AddUserUsersPostWithHttpInfo}{\code{DefaultApi$AddUserUsersPostWithHttpInfo()}}
\item \href{#method-DeleteUserUsersUidDelete}{\code{DefaultApi$DeleteUserUsersUidDelete()}}
\item \href{#method-DeleteUserUsersUidDeleteWithHttpInfo}{\code{DefaultApi$DeleteUserUsersUidDeleteWithHttpInfo()}}
\item \href{#method-FilesChangeFilesChangesPost}{\code{DefaultApi$FilesChangeFilesChangesPost()}}
\item \href{#method-FilesChangeFilesChangesPostWithHttpInfo}{\code{DefaultApi$FilesChangeFilesChangesPostWithHttpInfo()}}
\item \href{#method-GenerateDownloadSignedUrlV4FilesSignedGet}{\code{DefaultApi$GenerateDownloadSignedUrlV4FilesSignedGet()}}
\item \href{#method-GenerateDownloadSignedUrlV4FilesSignedGetWithHttpInfo}{\code{DefaultApi$GenerateDownloadSignedUrlV4FilesSignedGetWithHttpInfo()}}
\item \href{#method-GetNextflowEventNextflowEventsIdGet}{\code{DefaultApi$GetNextflowEventNextflowEventsIdGet()}}
\item \href{#method-GetNextflowEventNextflowEventsIdGetWithHttpInfo}{\code{DefaultApi$GetNextflowEventNextflowEventsIdGetWithHttpInfo()}}
\item \href{#method-GetUserUsersUidGet}{\code{DefaultApi$GetUserUsersUidGet()}}
\item \href{#method-GetUserUsersUidGetWithHttpInfo}{\code{DefaultApi$GetUserUsersUidGetWithHttpInfo()}}
\item \href{#method-ListFileEventsFilesChangesGet}{\code{DefaultApi$ListFileEventsFilesChangesGet()}}
\item \href{#method-ListFileEventsFilesChangesGetWithHttpInfo}{\code{DefaultApi$ListFileEventsFilesChangesGetWithHttpInfo()}}
\item \href{#method-ListNextflowEventsNextflowEventsGet}{\code{DefaultApi$ListNextflowEventsNextflowEventsGet()}}
\item \href{#method-ListNextflowEventsNextflowEventsGetWithHttpInfo}{\code{DefaultApi$ListNextflowEventsNextflowEventsGetWithHttpInfo()}}
\item \href{#method-ListUsersUsersGet}{\code{DefaultApi$ListUsersUsersGet()}}
\item \href{#method-ListUsersUsersGetWithHttpInfo}{\code{DefaultApi$ListUsersUsersGetWithHttpInfo()}}
\item \href{#method-clone}{\code{DefaultApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddNextflowEventNextflowEventsPost"></a>}}
\if{latex}{\out{\hypertarget{method-AddNextflowEventNextflowEventsPost}{}}}
\subsection{Method \code{AddNextflowEventNextflowEventsPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$AddNextflowEventNextflowEventsPost(nextflow.event.model, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddNextflowEventNextflowEventsPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AddNextflowEventNextflowEventsPostWithHttpInfo}{}}}
\subsection{Method \code{AddNextflowEventNextflowEventsPostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$AddNextflowEventNextflowEventsPostWithHttpInfo(
  nextflow.event.model,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddUserUsersPost"></a>}}
\if{latex}{\out{\hypertarget{method-AddUserUsersPost}{}}}
\subsection{Method \code{AddUserUsersPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$AddUserUsersPost(user.model, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddUserUsersPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-AddUserUsersPostWithHttpInfo}{}}}
\subsection{Method \code{AddUserUsersPostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$AddUserUsersPostWithHttpInfo(user.model, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteUserUsersUidDelete"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteUserUsersUidDelete}{}}}
\subsection{Method \code{DeleteUserUsersUidDelete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$DeleteUserUsersUidDelete(uid, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DeleteUserUsersUidDeleteWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DeleteUserUsersUidDeleteWithHttpInfo}{}}}
\subsection{Method \code{DeleteUserUsersUidDeleteWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$DeleteUserUsersUidDeleteWithHttpInfo(uid, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilesChangeFilesChangesPost"></a>}}
\if{latex}{\out{\hypertarget{method-FilesChangeFilesChangesPost}{}}}
\subsection{Method \code{FilesChangeFilesChangesPost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$FilesChangeFilesChangesPost(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FilesChangeFilesChangesPostWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-FilesChangeFilesChangesPostWithHttpInfo}{}}}
\subsection{Method \code{FilesChangeFilesChangesPostWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$FilesChangeFilesChangesPostWithHttpInfo(body, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenerateDownloadSignedUrlV4FilesSignedGet"></a>}}
\if{latex}{\out{\hypertarget{method-GenerateDownloadSignedUrlV4FilesSignedGet}{}}}
\subsection{Method \code{GenerateDownloadSignedUrlV4FilesSignedGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GenerateDownloadSignedUrlV4FilesSignedGet(name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GenerateDownloadSignedUrlV4FilesSignedGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GenerateDownloadSignedUrlV4FilesSignedGetWithHttpInfo}{}}}
\subsection{Method \code{GenerateDownloadSignedUrlV4FilesSignedGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GenerateDownloadSignedUrlV4FilesSignedGetWithHttpInfo(name, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetNextflowEventNextflowEventsIdGet"></a>}}
\if{latex}{\out{\hypertarget{method-GetNextflowEventNextflowEventsIdGet}{}}}
\subsection{Method \code{GetNextflowEventNextflowEventsIdGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetNextflowEventNextflowEventsIdGet(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetNextflowEventNextflowEventsIdGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetNextflowEventNextflowEventsIdGetWithHttpInfo}{}}}
\subsection{Method \code{GetNextflowEventNextflowEventsIdGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetNextflowEventNextflowEventsIdGetWithHttpInfo(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUserUsersUidGet"></a>}}
\if{latex}{\out{\hypertarget{method-GetUserUsersUidGet}{}}}
\subsection{Method \code{GetUserUsersUidGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetUserUsersUidGet(uid, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GetUserUsersUidGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GetUserUsersUidGetWithHttpInfo}{}}}
\subsection{Method \code{GetUserUsersUidGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$GetUserUsersUidGetWithHttpInfo(uid, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFileEventsFilesChangesGet"></a>}}
\if{latex}{\out{\hypertarget{method-ListFileEventsFilesChangesGet}{}}}
\subsection{Method \code{ListFileEventsFilesChangesGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$ListFileEventsFilesChangesGet(
  limit = 100,
  offset = 0,
  start.date = NULL,
  end.date = NULL,
  filename.regex = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListFileEventsFilesChangesGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ListFileEventsFilesChangesGetWithHttpInfo}{}}}
\subsection{Method \code{ListFileEventsFilesChangesGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$ListFileEventsFilesChangesGetWithHttpInfo(
  limit = 100,
  offset = 0,
  start.date = NULL,
  end.date = NULL,
  filename.regex = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListNextflowEventsNextflowEventsGet"></a>}}
\if{latex}{\out{\hypertarget{method-ListNextflowEventsNextflowEventsGet}{}}}
\subsection{Method \code{ListNextflowEventsNextflowEventsGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$ListNextflowEventsNextflowEventsGet(limit = 100, offset = 0, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListNextflowEventsNextflowEventsGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ListNextflowEventsNextflowEventsGetWithHttpInfo}{}}}
\subsection{Method \code{ListNextflowEventsNextflowEventsGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$ListNextflowEventsNextflowEventsGetWithHttpInfo(
  limit = 100,
  offset = 0,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUsersUsersGet"></a>}}
\if{latex}{\out{\hypertarget{method-ListUsersUsersGet}{}}}
\subsection{Method \code{ListUsersUsersGet()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$ListUsersUsersGet(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ListUsersUsersGetWithHttpInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ListUsersUsersGetWithHttpInfo}{}}}
\subsection{Method \code{ListUsersUsersGetWithHttpInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$ListUsersUsersGetWithHttpInfo(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DefaultApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
