% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nf_events.R
\name{get_nf_events}
\alias{get_nf_events}
\title{Get nextflow events}
\usage{
get_nf_events(limit = 100, offset = 0)
}
\arguments{
\item{limit}{A numeric vector of length 1 indicating the number of rows that must be returned.}

\item{offset}{TODO}
}
\value{
tibble
}
\description{
Get nextflow events
}
\examples{

 # Get the first 100 file changes
x <- get_nf_events(limit = 100)
x

}
