% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_from_dir.R
\name{matrix_from_dir}
\alias{matrix_from_dir}
\title{Title}
\usage{
matrix_from_dir(dir_path)
}
\arguments{
\item{dir_path}{Character vector indicating path of the directory with tsv files.}
}
\value{
matrix
}
\description{
Title
}
\examples{



}
